added feature of auth to other platforms by deeplink

+++ app/src/main/java/com/example/data/models/UserSessions.kt
@@ -13,7 +13,7 @@ data class UserSessionModel(
     @SerializedName("userId")
     val userId: Int,
     @SerializedName("sessionId")
-    val sessionId: Int,
+    val sessionId: Long,
     @SerializedName("sessionUid")
     val sessionUid: String,
     @SerializedName("sessionStart")
